Open: fix polling
For now, create the window and poll until the window's ID disappears. A more elegant method that relies on the event log seems possible (and is implemented as waitevent()), but doesn't actually work. Weirdly, the event log seems to omit del events for the windows we create (but not others, and not non-del events).